restructured the tests for policies
